reservationをまとめる
